expose models kernels likelihoods
